requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.2
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
isort==5.12.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.4.0
